Move mongo stuff into repo folder
